+++ source/Pages/Notification/index.jsx
@@ -3,8 +3,6 @@ import { HANDLER_TYPES, sendMessage } from '@background/Keyring';
 import ReactDOM from 'react-dom';
 import qs from 'query-string';
 import { theme } from '@ui';
-import { useDispatch } from 'react-redux';
-import { setAccountInfo } from '@redux/wallet';
 import store from '../../redux/store';
 import ProviderWrapper from '../../shared/ProviderWrapper';
 import Login from '../../Popup/Views/Login';
@@ -22,7 +20,6 @@ const NOTIFICATION_COMPONENTS = {
 
 const NotificationContainer = () => {
   const [loggedIn, setLoggedIn] = useState(null);
-  const dispatch = useDispatch();
 
   const { query } = qs.parseUrl(window.location.href);
 
@@ -41,13 +38,6 @@ const NotificationContainer = () => {
         setLoggedIn(false);
       }
     });
-
-    sendMessage({ type: HANDLER_TYPES.GET_STATE, params: {} },
-      (state) => {
-        if (state?.wallets?.length) {
-          dispatch(setAccountInfo(state.wallets[0]));
-        }
-      });
   }, []);
 
   const handleLogin = () => setLoggedIn(true);
